# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Películas y Categorías",
        "description": "Documentación de la API para gestionar películas y categorías",
        "version": "1.0"
    },
    "paths": {
...
        "/api/peliculas": {
            "get": {
                "tags": [
                    "Películas"
                ],
                "summary": "Lista de películas",
                "description": "Obtener todas las películas.",
                "operationId": "ec16899ee7eef044d44557f0a3b44f3a",
                "responses": {
                    "200": {
                        "description": "Lista de películas obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pelicula"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Películas"
                ],
                "summary": "Crear una nueva película",
                "operationId": "8dbd956256347ce5405bba95dafe496b",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion",
                                    "duracion",
                                    "rutaImagen",
                                    "clasificacion",
                                    "categoria_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "duracion": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "rutaImagen": {
                                        "type": "file"
                                    },
                                    "clasificacion": {
                                        "type": "integer"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Película creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Película creada correctamente."
                                        },
                                        "pelicula": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Los datos proporcionados no son válidos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/peliculas/{id}": {
            "get": {
                "tags": [
                    "Películas"
                ],
                "summary": "Obtener una película específica",
                "description": "Obtener una película por ID.",
                "operationId": "2b010267bd7d1e90b36fc44ed91ae71a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la película",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Película obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pelicula"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Películas"
                ],
                "summary": "Actualizar una película",
                "description": "Actualizar una película (PUT).",
                "operationId": "96ec1444feac4a8955b1a942e84916bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion",
                                    "duracion",
                                    "rutaImagen",
                                    "rutaLocalImagen",
                                    "clasificacion",
                                    "categoria_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Avatar"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Película de ciencia ficción."
                                    },
                                    "duracion": {
                                        "type": "number",
                                        "example": 2.5
                                    },
                                    "rutaImagen": {
                                        "type": "string",
                                        "example": "https://example.com/avatar.jpg"
                                    },
                                    "rutaLocalImagen": {
                                        "type": "string",
                                        "example": "/images/avatar.jpg"
                                    },
                                    "clasificacion": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Película actualizada con éxito"
                    },
                    "404": {
                        "description": "Película no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Películas"
                ],
                "summary": "Eliminar una película",
                "description": "Eliminar una película.",
                "operationId": "b82c498aa94c5562ad492f5428049e59",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Película eliminada con éxito"
                    },
                    "404": {
                        "description": "Película no encontrada"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Películas"
                ],
                "summary": "Actualizar parcialmente una película",
                "description": "Actualización parcial de una película (PATCH).",
                "operationId": "976c457b7cb985c88ab58426cac48f8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Avatar"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Película de ciencia ficción."
                                    },
                                    "duracion": {
                                        "type": "number",
                                        "example": 2.5
                                    },
                                    "rutaImagen": {
                                        "type": "string",
                                        "example": "https://example.com/avatar.jpg"
                                    },
                                    "rutaLocalImagen": {
                                        "type": "string",
                                        "example": "/images/avatar.jpg"
                                    },
                                    "clasificacion": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Película actualizada parcialmente con éxito"
                    },
                    "404": {
                        "description": "Película no encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Pelicula": {
                "title": "Pelicula",
                "description": "Modelo de una película",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Titanic"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Película romántica sobre el Titanic"
                    },
                    "categoria_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}